requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Django==5.1.8
django-cors-headers==4.7.0
django-tenants==3.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
kombu==5.5.3
mysqlclient==2.2.7
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
